+++ mycodo/config.py
@@ -8,7 +8,7 @@ import collections
 
 from datetime import timedelta
 
-MYCODO_VERSION = '5.0.8'
+MYCODO_VERSION = '5.0.9'
 ALEMBIC_VERSION = '059a47f950b8'
 
 LANGUAGES = {